Prepare new add-subscription lambda
