+++ gfa-backend/Cargo.toml
@@ -25,8 +25,8 @@ name = "notify"
 path = "src/notify/main.rs"
 
 [[bin]]
-name = "subscribe"
-path = "src/subscribe/main.rs"
+name = "add-subscription"
+path = "src/add-subscription/main.rs"
 
 [lib]
 name = "common"
