--- gfa-backend/src/subscribe/validate_request.rs
@@ -1,8 +1,0 @@
-use std::{error};
-use crate::subscription::Subscription;
-
-pub fn validate(subscription: &Subscription) -> Result<(), Box<dyn error::Error>> {
-    // TODO: validate email format
-    // TODO: validate location-id exists?
-    Ok(())
-}
